etl/extract/moisturecarbonservice/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
httpx==0.23.3
requests==2.28.1
kafka-python==2.0.2